Marvel sample ui refactor

+++ rosie/src/main/java/com/karumi/rosie/repository/PaginatedRosieRepository.java
@@ -86,13 +86,15 @@ public class PaginatedRosieRepository<K, V extends Identifiable<K>> extends Rosi
     for (PaginatedCacheDataSource<K, V> cacheDataSource : paginatedCacheDataSources) {
       values = cacheDataSource.getPage(page);
 
-      if (values != null) {
+      if (values != null && values.getItems() != null && !values.getItems().isEmpty()) {
         if (areValidValues(values, cacheDataSource)) {
           break;
         } else {
           cacheDataSource.deleteAll();
           values = null;
         }
+      } else {
+        values = null;
       }
     }
 
@@ -105,7 +107,7 @@ public class PaginatedRosieRepository<K, V extends Identifiable<K>> extends Rosi
     for (PaginatedReadableDataSource<K, V> readable : paginatedReadableDataSources) {
       values = readable.getPage(page);
 
-      if (values != null) {
+      if (values != null && values.getItems() != null && !values.getItems().isEmpty()) {
         break;
       }
     }
